Split BlobService and BlobManagementService in their own libs (with their own extention-methods.
